@@ -4,11 +4,10 @@ description: Starting point for a Serverpod server.
 # homepage: https://www.example.com
 
 environment:
-  sdk: '>=2.17.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
-  serverpod:
-    path: ../vendor/serverpod/packages/serverpod
+  serverpod: 1.2.4
 
 dev_dependencies:
-  lints: ^2.0.0
+  lints: ^3.0.0
