@@ -4,18 +4,28 @@
 // ignore_for_file: library_private_types_in_public_api
 // ignore_for_file: public_member_api_docs
 // ignore_for_file: implementation_imports
+// ignore_for_file: use_super_parameters
+// ignore_for_file: type_literal_in_constant_pattern
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'package:serverpod/serverpod.dart' as _i1;
 import 'dart:typed_data' as _i2;
+import 'package:serverpod_serialization/serverpod_serialization.dart';
 
-class ImageData extends _i1.TableRow {
-  ImageData({
+abstract class ImageData extends _i1.TableRow {
+  ImageData._({
     int? id,
     required this.pixels,
     required this.width,
     required this.height,
   }) : super(id);
+
+  factory ImageData({
+    int? id,
+    required _i2.ByteData pixels,
+    required int width,
+    required int height,
+  }) = _ImageDataImpl;
 
   factory ImageData.fromJson(
     Map<String, dynamic> jsonSerialization,
@@ -33,6 +43,8 @@ class ImageData extends _i1.TableRow {
 
   static final t = ImageDataTable();
 
+  static const db = ImageDataRepository._();
+
   _i2.ByteData pixels;
 
   int width;
@@ -40,18 +52,26 @@ class ImageData extends _i1.TableRow {
   int height;
 
   @override
-  String get tableName => 'image_data';
+  _i1.Table get table => t;
+
+  ImageData copyWith({
+    int? id,
+    _i2.ByteData? pixels,
+    int? width,
+    int? height,
+  });
   @override
   Map<String, dynamic> toJson() {
     return {
-      'id': id,
-      'pixels': pixels,
+      if (id != null) 'id': id,
+      'pixels': pixels.toJson(),
       'width': width,
       'height': height,
     };
   }
 
   @override
+  @Deprecated('Will be removed in 2.0.0')
   Map<String, dynamic> toJsonForDatabase() {
     return {
       'id': id,
@@ -64,14 +84,15 @@ class ImageData extends _i1.TableRow {
   @override
   Map<String, dynamic> allToJson() {
     return {
-      'id': id,
-      'pixels': pixels,
+      if (id != null) 'id': id,
+      'pixels': pixels.toJson(),
       'width': width,
       'height': height,
     };
   }
 
   @override
+  @Deprecated('Will be removed in 2.0.0')
   void setColumn(
     String columnName,
     value,
@@ -94,9 +115,10 @@ class ImageData extends _i1.TableRow {
     }
   }
 
+  @Deprecated('Will be removed in 2.0.0. Use: db.find instead.')
   static Future<List<ImageData>> find(
     _i1.Session session, {
-    ImageDataExpressionBuilder? where,
+    _i1.WhereExpressionBuilder<ImageDataTable>? where,
     int? limit,
     int? offset,
     _i1.Column? orderBy,
@@ -117,9 +139,10 @@ class ImageData extends _i1.TableRow {
     );
   }
 
+  @Deprecated('Will be removed in 2.0.0. Use: db.findRow instead.')
   static Future<ImageData?> findSingleRow(
     _i1.Session session, {
-    ImageDataExpressionBuilder? where,
+    _i1.WhereExpressionBuilder<ImageDataTable>? where,
     int? offset,
     _i1.Column? orderBy,
     bool orderDescending = false,
@@ -136,6 +159,7 @@ class ImageData extends _i1.TableRow {
     );
   }
 
+  @Deprecated('Will be removed in 2.0.0. Use: db.findById instead.')
   static Future<ImageData?> findById(
     _i1.Session session,
     int id,
@@ -143,9 +167,10 @@ class ImageData extends _i1.TableRow {
     return session.db.findById<ImageData>(id);
   }
 
+  @Deprecated('Will be removed in 2.0.0. Use: db.deleteWhere instead.')
   static Future<int> delete(
     _i1.Session session, {
-    required ImageDataExpressionBuilder where,
+    required _i1.WhereExpressionBuilder<ImageDataTable> where,
     _i1.Transaction? transaction,
   }) async {
     return session.db.delete<ImageData>(
@@ -154,6 +179,7 @@ class ImageData extends _i1.TableRow {
     );
   }
 
+  @Deprecated('Will be removed in 2.0.0. Use: db.deleteRow instead.')
   static Future<bool> deleteRow(
     _i1.Session session,
     ImageData row, {
@@ -165,6 +191,7 @@ class ImageData extends _i1.TableRow {
     );
   }
 
+  @Deprecated('Will be removed in 2.0.0. Use: db.update instead.')
   static Future<bool> update(
     _i1.Session session,
     ImageData row, {
@@ -176,6 +203,8 @@ class ImageData extends _i1.TableRow {
     );
   }
 
+  @Deprecated(
+      'Will be removed in 2.0.0. Use: db.insert instead. Important note: In db.insert, the object you pass in is no longer modified, instead a new copy with the added row is returned which contains the inserted id.')
   static Future<void> insert(
     _i1.Session session,
     ImageData row, {
@@ -187,9 +216,10 @@ class ImageData extends _i1.TableRow {
     );
   }
 
+  @Deprecated('Will be removed in 2.0.0. Use: db.count instead.')
   static Future<int> count(
     _i1.Session session, {
-    ImageDataExpressionBuilder? where,
+    _i1.WhereExpressionBuilder<ImageDataTable>? where,
     int? limit,
     bool useCache = true,
     _i1.Transaction? transaction,
@@ -201,20 +231,84 @@ class ImageData extends _i1.TableRow {
       transaction: transaction,
     );
   }
+
+  static ImageDataInclude include() {
+    return ImageDataInclude._();
+  }
+
+  static ImageDataIncludeList includeList({
+    _i1.WhereExpressionBuilder<ImageDataTable>? where,
+    int? limit,
+    int? offset,
+    _i1.OrderByBuilder<ImageDataTable>? orderBy,
+    bool orderDescending = false,
+    _i1.OrderByListBuilder<ImageDataTable>? orderByList,
+    ImageDataInclude? include,
+  }) {
+    return ImageDataIncludeList._(
+      where: where,
+      limit: limit,
+      offset: offset,
+      orderBy: orderBy?.call(ImageData.t),
+      orderDescending: orderDescending,
+      orderByList: orderByList?.call(ImageData.t),
+      include: include,
+    );
+  }
 }
 
-typedef ImageDataExpressionBuilder = _i1.Expression Function(ImageDataTable);
+class _Undefined {}
+
+class _ImageDataImpl extends ImageData {
+  _ImageDataImpl({
+    int? id,
+    required _i2.ByteData pixels,
+    required int width,
+    required int height,
+  }) : super._(
+          id: id,
+          pixels: pixels,
+          width: width,
+          height: height,
+        );
+
+  @override
+  ImageData copyWith({
+    Object? id = _Undefined,
+    _i2.ByteData? pixels,
+    int? width,
+    int? height,
+  }) {
+    return ImageData(
+      id: id is int? ? id : this.id,
+      pixels: pixels ?? this.pixels.clone(),
+      width: width ?? this.width,
+      height: height ?? this.height,
+    );
+  }
+}
 
 class ImageDataTable extends _i1.Table {
-  ImageDataTable() : super(tableName: 'image_data');
+  ImageDataTable({super.tableRelation}) : super(tableName: 'image_data') {
+    pixels = _i1.ColumnByteData(
+      'pixels',
+      this,
+    );
+    width = _i1.ColumnInt(
+      'width',
+      this,
+    );
+    height = _i1.ColumnInt(
+      'height',
+      this,
+    );
+  }
 
-  final id = _i1.ColumnInt('id');
+  late final _i1.ColumnByteData pixels;
 
-  final pixels = _i1.ColumnByteData('pixels');
+  late final _i1.ColumnInt width;
 
-  final width = _i1.ColumnInt('width');
-
-  final height = _i1.ColumnInt('height');
+  late final _i1.ColumnInt height;
 
   @override
   List<_i1.Column> get columns => [
@@ -227,3 +321,182 @@ class ImageDataTable extends _i1.Table {
 
 @Deprecated('Use ImageDataTable.t instead.')
 ImageDataTable tImageData = ImageDataTable();
+
+class ImageDataInclude extends _i1.IncludeObject {
+  ImageDataInclude._();
+
+  @override
+  Map<String, _i1.Include?> get includes => {};
+
+  @override
+  _i1.Table get table => ImageData.t;
+}
+
+class ImageDataIncludeList extends _i1.IncludeList {
+  ImageDataIncludeList._({
+    _i1.WhereExpressionBuilder<ImageDataTable>? where,
+    super.limit,
+    super.offset,
+    super.orderBy,
+    super.orderDescending,
+    super.orderByList,
+    super.include,
+  }) {
+    super.where = where?.call(ImageData.t);
+  }
+
+  @override
+  Map<String, _i1.Include?> get includes => include?.includes ?? {};
+
+  @override
+  _i1.Table get table => ImageData.t;
+}
+
+class ImageDataRepository {
+  const ImageDataRepository._();
+
+  Future<List<ImageData>> find(
+    _i1.Session session, {
+    _i1.WhereExpressionBuilder<ImageDataTable>? where,
+    int? limit,
+    int? offset,
+    _i1.OrderByBuilder<ImageDataTable>? orderBy,
+    bool orderDescending = false,
+    _i1.OrderByListBuilder<ImageDataTable>? orderByList,
+    _i1.Transaction? transaction,
+  }) async {
+    return session.dbNext.find<ImageData>(
+      where: where?.call(ImageData.t),
+      orderBy: orderBy?.call(ImageData.t),
+      orderByList: orderByList?.call(ImageData.t),
+      orderDescending: orderDescending,
+      limit: limit,
+      offset: offset,
+      transaction: transaction,
+    );
+  }
+
+  Future<ImageData?> findFirstRow(
+    _i1.Session session, {
+    _i1.WhereExpressionBuilder<ImageDataTable>? where,
+    int? offset,
+    _i1.OrderByBuilder<ImageDataTable>? orderBy,
+    bool orderDescending = false,
+    _i1.OrderByListBuilder<ImageDataTable>? orderByList,
+    _i1.Transaction? transaction,
+  }) async {
+    return session.dbNext.findFirstRow<ImageData>(
+      where: where?.call(ImageData.t),
+      orderBy: orderBy?.call(ImageData.t),
+      orderByList: orderByList?.call(ImageData.t),
+      orderDescending: orderDescending,
+      offset: offset,
+      transaction: transaction,
+    );
+  }
+
+  Future<ImageData?> findById(
+    _i1.Session session,
+    int id, {
+    _i1.Transaction? transaction,
+  }) async {
+    return session.dbNext.findById<ImageData>(
+      id,
+      transaction: transaction,
+    );
+  }
+
+  Future<List<ImageData>> insert(
+    _i1.Session session,
+    List<ImageData> rows, {
+    _i1.Transaction? transaction,
+  }) async {
+    return session.dbNext.insert<ImageData>(
+      rows,
+      transaction: transaction,
+    );
+  }
+
+  Future<ImageData> insertRow(
+    _i1.Session session,
+    ImageData row, {
+    _i1.Transaction? transaction,
+  }) async {
+    return session.dbNext.insertRow<ImageData>(
+      row,
+      transaction: transaction,
+    );
+  }
+
+  Future<List<ImageData>> update(
+    _i1.Session session,
+    List<ImageData> rows, {
+    _i1.ColumnSelections<ImageDataTable>? columns,
+    _i1.Transaction? transaction,
+  }) async {
+    return session.dbNext.update<ImageData>(
+      rows,
+      columns: columns?.call(ImageData.t),
+      transaction: transaction,
+    );
+  }
+
+  Future<ImageData> updateRow(
+    _i1.Session session,
+    ImageData row, {
+    _i1.ColumnSelections<ImageDataTable>? columns,
+    _i1.Transaction? transaction,
+  }) async {
+    return session.dbNext.updateRow<ImageData>(
+      row,
+      columns: columns?.call(ImageData.t),
+      transaction: transaction,
+    );
+  }
+
+  Future<List<int>> delete(
+    _i1.Session session,
+    List<ImageData> rows, {
+    _i1.Transaction? transaction,
+  }) async {
+    return session.dbNext.delete<ImageData>(
+      rows,
+      transaction: transaction,
+    );
+  }
+
+  Future<int> deleteRow(
+    _i1.Session session,
+    ImageData row, {
+    _i1.Transaction? transaction,
+  }) async {
+    return session.dbNext.deleteRow<ImageData>(
+      row,
+      transaction: transaction,
+    );
+  }
+
+  Future<List<int>> deleteWhere(
+    _i1.Session session, {
+    required _i1.WhereExpressionBuilder<ImageDataTable> where,
+    _i1.Transaction? transaction,
+  }) async {
+    return session.dbNext.deleteWhere<ImageData>(
+      where: where(ImageData.t),
+      transaction: transaction,
+    );
+  }
+
+  Future<int> count(
+    _i1.Session session, {
+    _i1.WhereExpressionBuilder<ImageDataTable>? where,
+    int? limit,
+    _i1.Transaction? transaction,
+  }) async {
+    return session.dbNext.count<ImageData>(
+      where: where?.call(ImageData.t),
+      limit: limit,
+      transaction: transaction,
+    );
+  }
+}
