@@ -4,15 +4,17 @@
 // ignore_for_file: library_private_types_in_public_api
 // ignore_for_file: public_member_api_docs
 // ignore_for_file: implementation_imports
+// ignore_for_file: use_super_parameters
+// ignore_for_file: type_literal_in_constant_pattern
 
 library protocol; // ignore_for_file: no_leading_underscores_for_library_prefixes
 
 import 'package:serverpod/serverpod.dart' as _i1;
-import 'image_data.dart' as _i2;
-import 'image_update.dart' as _i3;
-import 'package:serverpod/protocol.dart' as _i4;
+import 'package:serverpod/protocol.dart' as _i2;
+import 'image_data.dart' as _i3;
+import 'image_update.dart' as _i4;
 export 'image_data.dart';
-export 'image_update.dart'; // ignore_for_file: equal_keys_in_map
+export 'image_update.dart';
 
 class Protocol extends _i1.SerializationManagerServer {
   Protocol._();
@@ -23,6 +25,60 @@ class Protocol extends _i1.SerializationManagerServer {
 
   static final Protocol _instance = Protocol._();
 
+  static final List<_i2.TableDefinition> targetTableDefinitions = [
+    _i2.TableDefinition(
+      name: 'image_data',
+      dartName: 'ImageData',
+      schema: 'public',
+      module: 'pixorama',
+      columns: [
+        _i2.ColumnDefinition(
+          name: 'id',
+          columnType: _i2.ColumnType.integer,
+          isNullable: false,
+          dartType: 'int?',
+          columnDefault: 'nextval(\'image_data_id_seq\'::regclass)',
+        ),
+        _i2.ColumnDefinition(
+          name: 'pixels',
+          columnType: _i2.ColumnType.bytea,
+          isNullable: false,
+          dartType: 'dart:typed_data:ByteData',
+        ),
+        _i2.ColumnDefinition(
+          name: 'width',
+          columnType: _i2.ColumnType.integer,
+          isNullable: false,
+          dartType: 'int',
+        ),
+        _i2.ColumnDefinition(
+          name: 'height',
+          columnType: _i2.ColumnType.integer,
+          isNullable: false,
+          dartType: 'int',
+        ),
+      ],
+      foreignKeys: [],
+      indexes: [
+        _i2.IndexDefinition(
+          indexName: 'image_data_pkey',
+          tableSpace: null,
+          elements: [
+            _i2.IndexElementDefinition(
+              type: _i2.IndexElementDefinitionType.column,
+              definition: 'id',
+            )
+          ],
+          type: 'btree',
+          isUnique: true,
+          isPrimary: true,
+        )
+      ],
+      managed: true,
+    ),
+    ..._i2.Protocol.targetTableDefinitions,
+  ];
+
   @override
   T deserialize<T>(
     dynamic data, [
@@ -32,30 +88,30 @@ class Protocol extends _i1.SerializationManagerServer {
     if (customConstructors.containsKey(t)) {
       return customConstructors[t]!(data, this) as T;
     }
-    if (t == _i2.ImageData) {
-      return _i2.ImageData.fromJson(data, this) as T;
+    if (t == _i3.ImageData) {
+      return _i3.ImageData.fromJson(data, this) as T;
     }
-    if (t == _i3.ImageUpdate) {
-      return _i3.ImageUpdate.fromJson(data, this) as T;
+    if (t == _i4.ImageUpdate) {
+      return _i4.ImageUpdate.fromJson(data, this) as T;
     }
-    if (t == _i1.getType<_i2.ImageData?>()) {
-      return (data != null ? _i2.ImageData.fromJson(data, this) : null) as T;
+    if (t == _i1.getType<_i3.ImageData?>()) {
+      return (data != null ? _i3.ImageData.fromJson(data, this) : null) as T;
     }
-    if (t == _i1.getType<_i3.ImageUpdate?>()) {
-      return (data != null ? _i3.ImageUpdate.fromJson(data, this) : null) as T;
+    if (t == _i1.getType<_i4.ImageUpdate?>()) {
+      return (data != null ? _i4.ImageUpdate.fromJson(data, this) : null) as T;
     }
     try {
-      return _i4.Protocol().deserialize<T>(data, t);
+      return _i2.Protocol().deserialize<T>(data, t);
     } catch (_) {}
     return super.deserialize<T>(data, t);
   }
 
   @override
   String? getClassNameForObject(Object data) {
-    if (data is _i2.ImageData) {
+    if (data is _i3.ImageData) {
       return 'ImageData';
     }
-    if (data is _i3.ImageUpdate) {
+    if (data is _i4.ImageUpdate) {
       return 'ImageUpdate';
     }
     return super.getClassNameForObject(data);
@@ -64,10 +120,10 @@ class Protocol extends _i1.SerializationManagerServer {
   @override
   dynamic deserializeByClassName(Map<String, dynamic> data) {
     if (data['className'] == 'ImageData') {
-      return deserialize<_i2.ImageData>(data['data']);
+      return deserialize<_i3.ImageData>(data['data']);
     }
     if (data['className'] == 'ImageUpdate') {
-      return deserialize<_i3.ImageUpdate>(data['data']);
+      return deserialize<_i4.ImageUpdate>(data['data']);
     }
     return super.deserializeByClassName(data);
   }
@@ -75,15 +131,22 @@ class Protocol extends _i1.SerializationManagerServer {
   @override
   _i1.Table? getTableForType(Type t) {
     {
-      var table = _i4.Protocol().getTableForType(t);
+      var table = _i2.Protocol().getTableForType(t);
       if (table != null) {
         return table;
       }
     }
     switch (t) {
-      case _i2.ImageData:
-        return _i2.ImageData.t;
+      case _i3.ImageData:
+        return _i3.ImageData.t;
     }
     return null;
   }
+
+  @override
+  List<_i2.TableDefinition> getTargetTableDefinitions() =>
+      targetTableDefinitions;
+
+  @override
+  String getModuleName() => 'pixorama';
 }
